requirements.txt
git+https://github.com/dbca-wa/webtemplate2.git#egg=webtemplate_dbca
git+https://github.com/dbca-wa/dbca-utils@1.1.1#egg=dbca-utils
psycopg2-binary>=2.9.4
Django~=5.0.8
django-extensions==3.2.3
ipython==7.19.0
social-auth-app-django==2.1.0
git+https://github.com/dbca-wa/ledger_api_client.git#egg=ledger_api_client
git+https://github.com/scottp-dpaw/social-core.git@email_fix#egg=social-auth-core
coverage==4.3.1
coveralls==1.1
reportlab==4.2.2
crispy_bootstrap5==2024.2
django-braces>=1.8.1
django-datatables-view==1.13.0
django-reversion==5.0.12
django-reversion-compare==0.8.6
git+https://github.com/xzzy/django-preserialize.git#egg=django-preserialize
django-countries==7.6.1
git+https://github.com/Tivix/django-cron#egg=django-cron
django-dynamic-fixture==1.9.1
openpyxl==3.0.7
datapackage==1.15.2
python-dateutil==2.6.1
py4j==0.10.2.1
djangorestframework==3.15.2
djangorestframework-gis==1.0
djangorestframework-csv==3.0.2
pycountry==17.1.8
six>=1.10.0
django-ical==1.9.2
django-taggit==6.0.0
geojson==2.5.0
unicodecsv==0.14.1
PyPDF2==1.26.0
django-extra-fields==0.9
django-crispy-forms==2.3
django-reset-migrations==0.4.0
django-ckeditor==5.4.0
djangorestframework-datatables==0.4.0
xlwt==1.3.0
xlsxwriter==1.1.2
gunicorn==23.0.0
dj-static==0.0.6
Pillow==10.3.0
mixer==6.1.3
django-dirtyfields==1.9.3
python-docx==0.8.11
psutil==5.7.0
git+https://github.com/xzzy/django_site_queue.git#egg=django_site_queue
git+https://github.com/dbca-wa/django-media-serv.git#egg=django_media_serv
dj-database-url==0.5.0
phonenumbers==8.12.15
django-treebeard==4.7.1
# django-utils-six==2.0
git+https://github.com/dbca-wa/django5-utils-six.git#egg=django5_six
django-six==1.0.5
django-widget-tweaks==1.5.0
django-phonenumber-field==8.0.0
gevent==23.9.1
whitenoise==5.3.0
django-summernote==0.8.20.0
Babel==2.15.0
encoders==0.0.4
git+https://github.com/dbca-wa/appmonitor_client.git#egg=appmonitor_client
python-decouple>=3.8
numpy==1.26.3
setuptools==72.2.0
wheel==0.42.0
git+https://github.com/dbca-wa/wagov_utils
idna==3.7
sqlparse==0.5.0
sentry-sdk~=1.44
pytz==2024.1
python-decouple==3.8
